Move the endpoints object to its own file
