+++ index.js
@@ -12,292 +12,14 @@ var _httpStatus = require('http-status');
 
 var _httpStatus2 = _interopRequireDefault(_httpStatus);
 
+var _endponts = require('./endponts');
+
+var _endponts2 = _interopRequireDefault(_endponts);
+
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-function _toConsumableArray(arr) { if (Array.isArray(arr)) { for (var i = 0, arr2 = Array(arr.length); i < arr.length; i++) { arr2[i] = arr[i]; } return arr2; } else { return Array.from(arr); } }
-
-var defaultEndpointBehaviour = {
-
-  // Default failure method for all endpoint calls
-  onFailure: async function onFailure(response) {
-    return {
-      status: response.status,
-      reason: response.statusText,
-      body: await response.json()
-    };
-  },
-
-  // Default success method for all endpoint calls
-  onSuccess: async function onSuccess(request) {
-    return await request.response.json();
-  }
-};
-
-var WealthSimpleTradeEndpoints = {
-
-  /*
-   * The LOGIN endpoint intializes a new session for the given email and
-   * password set. If the login is successful, access and refresh tokens
-   * are returned in the headers. The access token is the key for invoking
-   * all other end points.
-   */
-  LOGIN: {
-    method: "POST",
-    url: "https://trade-service.wealthsimple.com/auth/login",
-    onSuccess: async function onSuccess(request) {
-      return {
-        tokens: {
-          access: request.response.headers.get('x-access-token'),
-          refresh: request.response.headers.get('x-refresh-token')
-        },
-
-        accountInfo: await request.response.json()
-      };
-    },
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * Generates a new set of access and refresh tokens.
-   */
-  REFRESH: {
-    method: "POST",
-    url: "https://trade-service.wealthsimple.com/auth/refresh",
-    onSuccess: async function onSuccess(request) {
-      return {
-        access: request.response.headers.get('x-access-token'),
-        refresh: request.response.headers.get('x-refresh-token')
-      };
-    },
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * Grabs all account ids in this WealthSimple Trade account.
-   */
-  ACCOUNT_IDS: {
-    method: "GET",
-    url: "https://trade-service.wealthsimple.com/account/list",
-    onSuccess: async function onSuccess(request) {
-      var data = await request.response.json();
-
-      // Collect all account ids registered under this WealthSimple Trade Account
-      return data.results.map(function (account) {
-        return account.id;
-      });
-    },
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * The LIST_ACCOUNT endpoint retrieves general metadata of the
-   * WealthSimple Trade account, including balances, account id, and
-   * more.
-   */
-  LIST_ACCOUNT: {
-    method: "GET",
-    url: "https://trade-service.wealthsimple.com/account/list",
-    onSuccess: defaultEndpointBehaviour.onSuccess,
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * The HISTORY_ACCOUNT endpoint provides historical snapshots of the
-   * WealthSimple account for a specified timeframe.
-   */
-  HISTORY_ACCOUNT: {
-    method: "GET",
-    url: "https://trade-service.wealthsimple.com/account/history/{0}?account_id={1}",
-    parameters: {
-      0: "interval",
-      1: "accountId"
-    },
-    onSuccess: defaultEndpointBehaviour.onSuccess,
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * All deposits under the WealthSimple Trade account
-   */
-  DEPOSITS: {
-    method: "GET",
-    url: "https://trade-service.wealthsimple.com/deposits",
-    onSuccess: async function onSuccess(request) {
-      var data = await request.response.json();
-      return data.results;
-    },
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * All linked bank accounts under the WealthSimple Trade account
-   */
-  BANK_ACCOUNTS: {
-    method: "GET",
-    url: "https://trade-service.wealthsimple.com/bank-accounts",
-    onSuccess: async function onSuccess(request) {
-      var data = await request.response.json();
-      return data.results;
-    },
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * Current WealthSimple Trade USD/CAD exchange rates
-   */
-  EXCHANGE_RATES: {
-    method: "GET",
-    url: "https://trade-service.wealthsimple.com/forex",
-    onSuccess: defaultEndpointBehaviour.onSuccess,
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * Grabs information about the security resembled by the ticker.
-   */
-  SECURITY: {
-    method: "GET",
-    url: "https://trade-service.wealthsimple.com/securities?query={0}",
-    parameters: {
-      0: "ticker"
-    },
-    onSuccess: async function onSuccess(request) {
-      var data = await request.response.json();
-
-      if (data.results.length === 0) {
-        return Promise.reject({
-          reason: 'Security does not exist'
-        });
-      }
-
-      return data.results[0];
-    },
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * Grab a page of orders (20 orders).
-   */
-  ORDERS_BY_PAGE: {
-    method: "GET",
-    url: "https://trade-service.wealthsimple.com/orders?offset={0}&account_id={1}",
-    parameters: {
-      0: "offset",
-      1: "accountId"
-    },
-    onSuccess: async function onSuccess(request) {
-      var data = await request.response.json();
-      return {
-        total: data.total,
-        orders: data.results
-      };
-    },
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * Pull all orders (filled, cancelled, pending) for the specified account under
-   * the WealthSimple Trade account.
-   */
-  ALL_ORDERS: {
-    method: "GET",
-    url: "https://trade-service.wealthsimple.com/orders?account_id={0}",
-    parameters: {
-      0: "accountId"
-    },
-    onSuccess: async function onSuccess(request, tokens) {
-      var data = await request.response.json();
-      var pages = Math.ceil(data.total / ORDERS_PER_PAGE);
-      var orders = data.results;
-
-      if (pages > 1) {
-
-        // Query the rest of the pages
-        for (var page = 2; page <= pages; page++) {
-          var tmp = await wealthsimple.getOrdersByPage(tokens, request.arguments.accountId, page);
-          orders.push.apply(orders, _toConsumableArray(tmp.orders));
-        }
-      }
-
-      return {
-        total: orders.length,
-        orders: orders
-      };
-    },
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * Filters orders by status and ticker.
-   */
-  FILTERED_ORDERS: {
-    method: "GET",
-    url: "https://trade-service.wealthsimple.com/orders?account_id={0}",
-    parameters: {
-      0: "accountId"
-    },
-    onSuccess: async function onSuccess(request, tokens) {
-      var data = await request.response.json();
-      var pages = Math.ceil(data.total / ORDERS_PER_PAGE);
-
-      // The ticker symbol restricts the pending orders to a specific security
-      var pendingFilter = request.arguments.ticker ? function (order) {
-        return order.symbol === request.arguments.ticker && order.status === request.arguments.status;
-      } : function (order) {
-        return order.status === request.arguments.status;
-      };
-
-      var orders = data.results.filter(pendingFilter);
-      if (pages > 1) {
-
-        // Check all other pages for pending orders
-        for (var page = 2; page <= pages; page++) {
-          var tmp = await wealthsimple.getOrdersByPage(tokens, request.arguments.accountId, page);
-          orders.push.apply(orders, _toConsumableArray(tmp.orders.filter(pendingFilter)));
-        }
-      }
-
-      return {
-        total: orders.length,
-        orders: orders
-      };
-    },
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * Cancels a specific order by its id.
-   */
-  CANCEL_ORDER: {
-    method: "DELETE",
-    url: "https://trade-service.wealthsimple.com/orders/{0}",
-    parameters: {
-      0: "orderId"
-    },
-    onSuccess: async function onSuccess(request) {
-      return {
-        order: request.arguments.orderId,
-        response: await request.response.json()
-      };
-    },
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * Places an order for a security.
-   */
-  PLACE_ORDER: {
-    method: "POST",
-    url: "https://trade-service.wealthsimple.com/orders?account_id={0}",
-    parameters: {
-      0: "accountId"
-    },
-    onSuccess: defaultEndpointBehaviour.onSuccess,
-    onFailure: defaultEndpointBehaviour.onFailure
-  }
-
-  // The maximum number of orders retrieved by the /orders API.
-};var ORDERS_PER_PAGE = 20;
+// The maximum number of orders retrieved by the /orders API.
+var ORDERS_PER_PAGE = 20;
 var isSuccessfulRequest = function isSuccessfulRequest(code) {
   return httpSuccessCodes.includes(code);
 };
@@ -397,7 +119,7 @@ var wealthsimple = {
    * @param {*} password The password of the account
    */
   login: async function login(email, password) {
-    return handleRequest(WealthSimpleTradeEndpoints.LOGIN, { email: email, password: password });
+    return handleRequest(_endponts2.default.LOGIN, { email: email, password: password });
   },
 
   /**
@@ -406,7 +128,7 @@ var wealthsimple = {
    * @param {*} tokens The access and refresh tokens returned by a successful login.
    */
   refresh: async function refresh(tokens) {
-    return handleRequest(WealthSimpleTradeEndpoints.REFRESH, { refresh_token: tokens.refresh }, tokens);
+    return handleRequest(_endponts2.default.REFRESH, { refresh_token: tokens.refresh }, tokens);
   },
 
   /**
@@ -415,7 +137,7 @@ var wealthsimple = {
    * @param {*} tokens The access and refresh tokens returned by a successful login.
    */
   getAccounts: async function getAccounts(tokens) {
-    return handleRequest(WealthSimpleTradeEndpoints.ACCOUNT_IDS, {}, tokens);
+    return handleRequest(_endponts2.default.ACCOUNT_IDS, {}, tokens);
   },
 
   /**
@@ -424,7 +146,7 @@ var wealthsimple = {
    * @param {*} tokens The access and refresh tokens returned by a successful login.
    */
   getAccountData: async function getAccountData(tokens) {
-    return handleRequest(WealthSimpleTradeEndpoints.LIST_ACCOUNT, {}, tokens);
+    return handleRequest(_endponts2.default.LIST_ACCOUNT, {}, tokens);
   },
 
   /**
@@ -435,7 +157,7 @@ var wealthsimple = {
    * @param {*} accountId The account to query
    */
   getHistory: async function getHistory(tokens, interval, accountId) {
-    return handleRequest(WealthSimpleTradeEndpoints.HISTORY_ACCOUNT, { interval: interval, accountId: accountId }, tokens);
+    return handleRequest(_endponts2.default.HISTORY_ACCOUNT, { interval: interval, accountId: accountId }, tokens);
   },
 
   /**
@@ -444,7 +166,7 @@ var wealthsimple = {
    * @param {*} tokens The access and refresh tokens returned by a successful login.
    */
   getBankAccounts: async function getBankAccounts(tokens) {
-    return handleRequest(WealthSimpleTradeEndpoints.BANK_ACCOUNTS, {}, tokens);
+    return handleRequest(_endponts2.default.BANK_ACCOUNTS, {}, tokens);
   },
 
   /**
@@ -453,7 +175,7 @@ var wealthsimple = {
    * @param {*} tokens The access and refresh tokens returned by a successful login.
    */
   getDeposits: async function getDeposits(tokens) {
-    return handleRequest(WealthSimpleTradeEndpoints.DEPOSITS, {}, tokens);
+    return handleRequest(_endponts2.default.DEPOSITS, {}, tokens);
   },
 
   /**
@@ -463,7 +185,7 @@ var wealthsimple = {
    * @param {*} tokens The access and refresh tokens returned by a successful login.
    */
   getExchangeRates: async function getExchangeRates(tokens) {
-    return handleRequest(WealthSimpleTradeEndpoints.EXCHANGE_RATES, {}, tokens);
+    return handleRequest(_endponts2.default.EXCHANGE_RATES, {}, tokens);
   },
 
   /**
@@ -471,7 +193,7 @@ var wealthsimple = {
    * account id.
    */
   getOrdersByPage: async function getOrdersByPage(tokens, accountId, page) {
-    return handleRequest(WealthSimpleTradeEndpoints.ORDERS_BY_PAGE, {
+    return handleRequest(_endponts2.default.ORDERS_BY_PAGE, {
       offset: (page - 1) * ORDERS_PER_PAGE,
       accountId: accountId
     }, tokens);
@@ -484,7 +206,7 @@ var wealthsimple = {
    * @param {*} accountId The specific account in the WealthSimple Trade account
    */
   getOrders: async function getOrders(tokens, accountId) {
-    return handleRequest(WealthSimpleTradeEndpoints.ALL_ORDERS, {
+    return handleRequest(_endponts2.default.ALL_ORDERS, {
       offset: 0,
       accountId: accountId
     }, tokens);
@@ -498,7 +220,7 @@ var wealthsimple = {
    * @param {*} ticker (optional) The security symbol
    */
   getPendingOrders: async function getPendingOrders(tokens, accountId, ticker) {
-    return handleRequest(WealthSimpleTradeEndpoints.FILTERED_ORDERS, {
+    return handleRequest(_endponts2.default.FILTERED_ORDERS, {
       accountId: accountId,
       ticker: ticker,
       status: 'submitted'
@@ -513,7 +235,7 @@ var wealthsimple = {
    * @param {*} ticker (optional) The security symbol
    */
   getFilledOrders: async function getFilledOrders(tokens, accountId, ticker) {
-    return handleRequest(WealthSimpleTradeEndpoints.FILTERED_ORDERS, {
+    return handleRequest(_endponts2.default.FILTERED_ORDERS, {
       accountId: accountId,
       ticker: ticker,
       status: 'posted'
@@ -528,7 +250,7 @@ var wealthsimple = {
    * @param {*} ticker (optional) The security symbol
    */
   getCancelledOrders: async function getCancelledOrders(tokens, accountId, ticker) {
-    return handleRequest(WealthSimpleTradeEndpoints.FILTERED_ORDERS, {
+    return handleRequest(_endponts2.default.FILTERED_ORDERS, {
       accountId: accountId,
       ticker: ticker,
       status: 'cancelled'
@@ -542,7 +264,7 @@ var wealthsimple = {
    * @param {*} orderId The pending order to cancel
    */
   cancelOrder: async function cancelOrder(tokens, orderId) {
-    return handleRequest(WealthSimpleTradeEndpoints.CANCEL_ORDER, { orderId: orderId }, tokens);
+    return handleRequest(_endponts2.default.CANCEL_ORDER, { orderId: orderId }, tokens);
   },
 
   /**
@@ -565,7 +287,7 @@ var wealthsimple = {
    * @param {*} ticker The security symbol
    */
   getSecurity: async function getSecurity(tokens, ticker) {
-    return handleRequest(WealthSimpleTradeEndpoints.SECURITY, { ticker: ticker }, tokens);
+    return handleRequest(_endponts2.default.SECURITY, { ticker: ticker }, tokens);
   },
 
   /**
@@ -578,7 +300,7 @@ var wealthsimple = {
    * @param {*} quantity The number of securities to purchase
    */
   placeLimitBuy: async function placeLimitBuy(tokens, accountId, ticker, limit, quantity) {
-    return handleRequest(WealthSimpleTradeEndpoints.PLACE_ORDER, {
+    return handleRequest(_endponts2.default.PLACE_ORDER, {
       accountId: accountId,
       security_id: (await wealthsimple.getSecurity(tokens, ticker)).id,
       limit_price: limit,
@@ -599,7 +321,7 @@ var wealthsimple = {
    * @param {*} quantity The number of securities to sell
    */
   placeLimitSell: async function placeLimitSell(tokens, accountId, ticker, limit, quantity) {
-    return handleRequest(WealthSimpleTradeEndpoints.PLACE_ORDER, {
+    return handleRequest(_endponts2.default.PLACE_ORDER, {
       accountId: accountId,
       security_id: (await wealthsimple.getSecurity(tokens, ticker)).id,
       limit_price: limit,
