+++ src/index.js
@@ -1,280 +1,6 @@
 import fetch from 'node-fetch';
 import status from 'http-status';
-
-const defaultEndpointBehaviour = {
-
-  // Default failure method for all endpoint calls
-  onFailure: async (response) => {
-    return {
-      status: response.status,
-      reason: response.statusText,
-      body: await response.json()
-    }
-  },
-
-  // Default success method for all endpoint calls
-  onSuccess: async (request) => await request.response.json()
-}
-
-const WealthSimpleTradeEndpoints = {
-
-  /*
-   * The LOGIN endpoint intializes a new session for the given email and
-   * password set. If the login is successful, access and refresh tokens
-   * are returned in the headers. The access token is the key for invoking
-   * all other end points.
-   */
-  LOGIN: {
-    method: "POST",
-    url: "https://trade-service.wealthsimple.com/auth/login",
-    onSuccess: async (request) => {
-      return {
-        tokens: {
-          access: request.response.headers.get('x-access-token'),
-          refresh: request.response.headers.get('x-refresh-token')
-        },
-
-        accountInfo: await request.response.json()
-      };
-    },
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * Generates a new set of access and refresh tokens.
-   */
-  REFRESH: {
-    method: "POST",
-    url: "https://trade-service.wealthsimple.com/auth/refresh",
-    onSuccess: async (request) => {
-      return {
-        access: request.response.headers.get('x-access-token'),
-        refresh: request.response.headers.get('x-refresh-token')
-      }
-    },
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * Grabs all account ids in this WealthSimple Trade account.
-   */
-  ACCOUNT_IDS: {
-    method: "GET",
-    url: "https://trade-service.wealthsimple.com/account/list",
-    onSuccess: async (request) => {
-      const data = await request.response.json();
-
-      // Collect all account ids registered under this WealthSimple Trade Account
-      return data.results.map(account => account.id);
-    },
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * The LIST_ACCOUNT endpoint retrieves general metadata of the
-   * WealthSimple Trade account, including balances, account id, and
-   * more.
-   */
-  LIST_ACCOUNT: {
-    method: "GET",
-    url: "https://trade-service.wealthsimple.com/account/list",
-    onSuccess: defaultEndpointBehaviour.onSuccess,
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * The HISTORY_ACCOUNT endpoint provides historical snapshots of the
-   * WealthSimple account for a specified timeframe.
-   */
-  HISTORY_ACCOUNT: {
-    method: "GET",
-    url: "https://trade-service.wealthsimple.com/account/history/{0}?account_id={1}",
-    parameters: {
-      0: "interval",
-      1: "accountId"
-    },
-    onSuccess: defaultEndpointBehaviour.onSuccess,
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * All deposits under the WealthSimple Trade account
-   */
-  DEPOSITS: {
-    method: "GET",
-    url: "https://trade-service.wealthsimple.com/deposits",
-    onSuccess: async (request) => {
-      const data = await request.response.json();
-      return data.results;
-    },
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * All linked bank accounts under the WealthSimple Trade account
-   */
-  BANK_ACCOUNTS: {
-    method: "GET",
-    url: "https://trade-service.wealthsimple.com/bank-accounts",
-    onSuccess: async (request) => {
-      const data = await request.response.json();
-      return data.results;
-    },
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * Current WealthSimple Trade USD/CAD exchange rates
-   */
-  EXCHANGE_RATES : {
-    method: "GET",
-    url: "https://trade-service.wealthsimple.com/forex",
-    onSuccess: defaultEndpointBehaviour.onSuccess,
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * Grabs information about the security resembled by the ticker.
-   */
-  SECURITY: {
-    method: "GET",
-    url: "https://trade-service.wealthsimple.com/securities?query={0}",
-    parameters: {
-      0: "ticker"
-    },
-    onSuccess: async (request) => {
-      let data = await request.response.json();
-
-      if (data.results.length === 0) {
-        return Promise.reject({
-          reason: `Security does not exist`
-        });
-      }
-
-      return data.results[0];
-    },
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * Grab a page of orders (20 orders).
-   */
-  ORDERS_BY_PAGE: {
-    method: "GET",
-    url: "https://trade-service.wealthsimple.com/orders?offset={0}&account_id={1}",
-    parameters: {
-      0: "offset",
-      1: "accountId",
-    },
-    onSuccess: async (request) => {
-      const data = await request.response.json();
-      return {
-        total: data.total,
-        orders: data.results
-      }
-    },
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * Pull all orders (filled, cancelled, pending) for the specified account under
-   * the WealthSimple Trade account.
-   */
-  ALL_ORDERS: {
-    method: "GET",
-    url: "https://trade-service.wealthsimple.com/orders?account_id={0}",
-    parameters: {
-      0: "accountId",
-    },
-    onSuccess: async (request, tokens) => {
-      const data = await request.response.json();
-      const pages = Math.ceil(data.total / ORDERS_PER_PAGE);
-      let orders = data.results;
-
-      if (pages > 1) {
-
-        // Query the rest of the pages
-        for (let page = 2; page <= pages; page++) {
-          let tmp = await wealthsimple.getOrdersByPage(tokens, request.arguments.accountId, page);
-          orders.push(...tmp.orders)
-        }
-      }
-
-      return {
-        total: orders.length,
-        orders
-      };
-    },
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * Filters orders by status and ticker.
-   */
-  FILTERED_ORDERS: {
-    method: "GET",
-    url: "https://trade-service.wealthsimple.com/orders?account_id={0}",
-    parameters: {
-      0: "accountId"
-    },
-    onSuccess: async (request, tokens) => {
-      const data = await request.response.json();
-      const pages = Math.ceil(data.total / ORDERS_PER_PAGE);
-
-      // The ticker symbol restricts the pending orders to a specific security
-      let pendingFilter = (request.arguments.ticker) ?
-                          order => order.symbol === request.arguments.ticker && order.status === request.arguments.status:
-                          order => order.status === request.arguments.status;
-
-      let orders = data.results.filter(pendingFilter);
-      if (pages > 1) {
-
-        // Check all other pages for pending orders
-        for (let page = 2; page <= pages; page++) {
-          let tmp = await wealthsimple.getOrdersByPage(tokens, request.arguments.accountId, page);
-          orders.push(...tmp.orders.filter(pendingFilter))
-        }
-      }
-
-      return {
-        total: orders.length,
-        orders
-      };
-    },
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * Cancels a specific order by its id.
-   */
-  CANCEL_ORDER: {
-    method: "DELETE",
-    url: "https://trade-service.wealthsimple.com/orders/{0}",
-    parameters: {
-      0: "orderId"
-    },
-    onSuccess: async (request) => {
-      return {
-        order: request.arguments.orderId,
-        response: await request.response.json()
-      }
-    },
-    onFailure: defaultEndpointBehaviour.onFailure
-  },
-
-  /*
-   * Places an order for a security.
-   */
-  PLACE_ORDER: {
-    method: "POST",
-    url: "https://trade-service.wealthsimple.com/orders?account_id={0}",
-    parameters: {
-      0: "accountId"
-    },
-    onSuccess: defaultEndpointBehaviour.onSuccess,
-    onFailure: defaultEndpointBehaviour.onFailure
-  }
-}
+import endpoints from './endponts';
 
 // The maximum number of orders retrieved by the /orders API.
 const ORDERS_PER_PAGE = 20;
@@ -377,7 +103,7 @@ const wealthsimple = {
    * @param {*} password The password of the account
    */
   login: async (email, password) =>
-    handleRequest(WealthSimpleTradeEndpoints.LOGIN, { email, password }),
+    handleRequest(endpoints.LOGIN, { email, password }),
 
   /**
    * Generates a new set of access and refresh tokens.
@@ -385,7 +111,7 @@ const wealthsimple = {
    * @param {*} tokens The access and refresh tokens returned by a successful login.
    */
   refresh: async (tokens) =>
-    handleRequest(WealthSimpleTradeEndpoints.REFRESH, { refresh_token: tokens.refresh}, tokens),
+    handleRequest(endpoints.REFRESH, { refresh_token: tokens.refresh}, tokens),
 
   /**
    * Retrieves all account ids open under this WealthSimple Trade account.
@@ -393,7 +119,7 @@ const wealthsimple = {
    * @param {*} tokens The access and refresh tokens returned by a successful login.
    */
   getAccounts: async (tokens) =>
-    handleRequest(WealthSimpleTradeEndpoints.ACCOUNT_IDS, {}, tokens),
+    handleRequest(endpoints.ACCOUNT_IDS, {}, tokens),
 
   /**
    * Retrieves the top-level data of the account, including account id, account types, account values, and more.
@@ -401,7 +127,7 @@ const wealthsimple = {
    * @param {*} tokens The access and refresh tokens returned by a successful login.
    */
   getAccountData: async (tokens) =>
-    handleRequest(WealthSimpleTradeEndpoints.LIST_ACCOUNT, {}, tokens),
+    handleRequest(endpoints.LIST_ACCOUNT, {}, tokens),
 
   /**
    * Query the history of the account within a certain time interval.
@@ -411,7 +137,7 @@ const wealthsimple = {
    * @param {*} accountId The account to query
    */
   getHistory: async (tokens, interval, accountId) =>
-    handleRequest(WealthSimpleTradeEndpoints.HISTORY_ACCOUNT, { interval, accountId }, tokens),
+    handleRequest(endpoints.HISTORY_ACCOUNT, { interval, accountId }, tokens),
 
   /**
    * Retains all bank accounts linked to the WealthSimple Trade account.
@@ -419,7 +145,7 @@ const wealthsimple = {
    * @param {*} tokens The access and refresh tokens returned by a successful login.
    */
   getBankAccounts: async (tokens) =>
-    handleRequest(WealthSimpleTradeEndpoints.BANK_ACCOUNTS, {}, tokens),
+    handleRequest(endpoints.BANK_ACCOUNTS, {}, tokens),
 
   /**
    * Grab all deposit records on the WealthSimple Trade account.
@@ -427,7 +153,7 @@ const wealthsimple = {
    * @param {*} tokens The access and refresh tokens returned by a successful login.
    */
   getDeposits: async (tokens) =>
-    handleRequest(WealthSimpleTradeEndpoints.DEPOSITS, {}, tokens),
+    handleRequest(endpoints.DEPOSITS, {}, tokens),
 
   /**
    * A snapshots of the current USD/CAD exchange rates on the WealthSimple Trade
@@ -436,14 +162,14 @@ const wealthsimple = {
    * @param {*} tokens The access and refresh tokens returned by a successful login.
    */
   getExchangeRates: async (tokens) =>
-    handleRequest(WealthSimpleTradeEndpoints.EXCHANGE_RATES, {}, tokens),
+    handleRequest(endpoints.EXCHANGE_RATES, {}, tokens),
 
   /**
    * Collects orders (filled, pending, cancelled) for the provided page and
    * account id.
    */
   getOrdersByPage: async (tokens, accountId, page) =>
-    handleRequest(WealthSimpleTradeEndpoints.ORDERS_BY_PAGE, {
+    handleRequest(endpoints.ORDERS_BY_PAGE, {
       offset: (page - 1) * ORDERS_PER_PAGE,
       accountId
     }, tokens),
@@ -455,7 +181,7 @@ const wealthsimple = {
    * @param {*} accountId The specific account in the WealthSimple Trade account
    */
   getOrders: async (tokens, accountId) =>
-    handleRequest(WealthSimpleTradeEndpoints.ALL_ORDERS, {
+    handleRequest(endpoints.ALL_ORDERS, {
       offset: 0,
       accountId
     }, tokens),
@@ -468,7 +194,7 @@ const wealthsimple = {
    * @param {*} ticker (optional) The security symbol
    */
   getPendingOrders: async (tokens, accountId, ticker) =>
-    handleRequest(WealthSimpleTradeEndpoints.FILTERED_ORDERS, {
+    handleRequest(endpoints.FILTERED_ORDERS, {
       accountId,
       ticker,
       status: 'submitted'
@@ -482,7 +208,7 @@ const wealthsimple = {
    * @param {*} ticker (optional) The security symbol
    */
   getFilledOrders: async (tokens, accountId, ticker) =>
-    handleRequest(WealthSimpleTradeEndpoints.FILTERED_ORDERS, {
+    handleRequest(endpoints.FILTERED_ORDERS, {
       accountId,
       ticker,
       status: 'posted'
@@ -496,7 +222,7 @@ const wealthsimple = {
    * @param {*} ticker (optional) The security symbol
    */
   getCancelledOrders: async (tokens, accountId, ticker) =>
-    handleRequest(WealthSimpleTradeEndpoints.FILTERED_ORDERS, {
+    handleRequest(endpoints.FILTERED_ORDERS, {
       accountId,
       ticker,
       status: 'cancelled'
@@ -509,7 +235,7 @@ const wealthsimple = {
    * @param {*} orderId The pending order to cancel
    */
   cancelOrder: async (tokens, orderId) =>
-    handleRequest(WealthSimpleTradeEndpoints.CANCEL_ORDER, { orderId }, tokens),
+    handleRequest(endpoints.CANCEL_ORDER, { orderId }, tokens),
 
   /**
    * Cancels all pending orders under the WealthSimple Trade Account.
@@ -529,7 +255,7 @@ const wealthsimple = {
    * @param {*} ticker The security symbol
    */
   getSecurity: async (tokens, ticker) =>
-    handleRequest(WealthSimpleTradeEndpoints.SECURITY, { ticker }, tokens),
+    handleRequest(endpoints.SECURITY, { ticker }, tokens),
 
   /**
    * Limit buy a security through the WealthSimple Trade application.
@@ -541,7 +267,7 @@ const wealthsimple = {
    * @param {*} quantity The number of securities to purchase
    */
   placeLimitBuy: async(tokens, accountId, ticker, limit, quantity) =>
-    handleRequest(WealthSimpleTradeEndpoints.PLACE_ORDER, {
+    handleRequest(endpoints.PLACE_ORDER, {
       accountId,
       security_id: (await wealthsimple.getSecurity(tokens, ticker)).id,
       limit_price: limit,
@@ -561,7 +287,7 @@ const wealthsimple = {
    * @param {*} quantity The number of securities to sell
    */
   placeLimitSell: async (tokens, accountId, ticker, limit, quantity) =>
-    handleRequest(WealthSimpleTradeEndpoints.PLACE_ORDER, {
+    handleRequest(endpoints.PLACE_ORDER, {
       accountId,
       security_id: (await wealthsimple.getSecurity(tokens, ticker)).id,
       limit_price: limit,
